requirements.txt
mysql.connector == 2.2.9
matplotlib==3.8.4
numpy==1.26.4
tensorflow==2.16.1
keras== 3.2.0
opencv-python==4.9.0.80
pyqt5 == 5.15.9


